Исправление ошибок ts

+++ src/controlles/ChatsController.ts
@@ -1,17 +1,19 @@
+/* eslint-disable no-restricted-globals */
+/* eslint-disable camelcase */
+/* eslint-disable no-alert */
+/* eslint-disable no-undef */
 import ChatsApi from '../api/ChatsApi';
-import UsersApi from '../api/UsersApi';
-import { TOptionsData } from '../classes/HTTPTransport';
 import Store from '../classes/Store';
-import { searchObjInArray } from '../utils/object_utils';
 import BaseController from './BaseController';
 import MessageController from './MessageController';
 
 class ChatsController extends BaseController {
-    public async getChats() {
+    public async getChats(): Promise<void> {
         try {
+            if (this.router?._currentRoute?._pathname !== '/messenger') return;
             const { status, response } = await ChatsApi.getChats();
             if (status === 200) {
-                this.store.set('chats', JSON.parse(response));
+                Store.set('chats', JSON.parse(response));
             } else if (status === 500) {
                 this.router.go('/500');
             } else {
@@ -22,7 +24,7 @@ class ChatsController extends BaseController {
         }
     }
 
-    public async createChat(title) {
+    public async createChat(title: string): Promise<boolean | number> {
         if (!title) return false;
         try {
             const { status, response } = await ChatsApi.createChat(title);
@@ -31,21 +33,22 @@ class ChatsController extends BaseController {
                 await this.getChats();
                 MessageController.changeCurrentChat(chatId);
                 return chatId;
-            } else if (status === 500) {
+            } if (status === 500) {
                 this.router.go('/500');
                 return false;
-            } else {
-                alert(JSON.parse(response).reason ?? 'Ошибочный запрос');
-                return false;
             }
+            alert(JSON.parse(response).reason ?? 'Ошибочный запрос');
+            return false;
         } catch (e) {
             console.log(e);
+            return false;
         }
     }
 
-    public async deleteChats() {
+    public async deleteChats(): Promise<void> {
         try {
             const chatId = this?.store?.getState()?.currentChat?.chat?.id;
+            if (typeof chatId !== 'number') return;
             const { status, response } = await ChatsApi.deleteChat({ chatId });
             if (status === 200) {
                 this.getChats();
@@ -66,37 +69,40 @@ class ChatsController extends BaseController {
         }
     }
 
-    public async addUser(id: number | string, user: number | string) {
+    public async addUser(id: number, user: number): Promise<boolean> {
         if (!id || !user) return false;
         try {
             const { status, response } = await ChatsApi.addUsers(id, [user]);
             if (status === 200) {
                 return true;
-            } else if (status === 500) {
+            } if (status === 500) {
                 this.router.go('/500');
                 return false;
-            } else {
-                alert(JSON.parse(response).reason ?? 'Ошибочный запрос');
-                return false;
             }
+            alert(JSON.parse(response).reason ?? 'Ошибочный запрос');
+            return false;
         } catch (e) {
             console.log(e);
         }
+        return false;
     }
 
-    public async addNewChatUser({ display_name, login, id }) {
+    public async addNewChatUser(user: Record<string, string | number>): Promise<boolean | void> {
+        const { display_name, login, id } = user;
         let chat = this?.store?.getState()?.currentChat?.chat?.id;
+        if (!confirm(`Вы хотите ${chat ? 'добавить в текущий чат ' : 'создать новый чат с '}${login}`)) {
+            return;
+        }
         if (!chat) {
             const title = display_name ?? login;
-            chat = await this.createChat(title);
+            chat = await this.createChat(String(title));
+            return;
         }
         if (!chat) return;
-        console.log(chat);
-
-        const result = await this.addUser(chat, id);
+        const result = await this.addUser(Number(chat), Number(id));
+        // eslint-disable-next-line consistent-return
         return result;
     }
-
 }
 
 export default new ChatsController();
